Check download categories in downloadAheadChapters
